fix: move endpoint and small front-end change
