--- frontend/src/components/SeedlotReviewSteps/Collection/constants.ts
@@ -1,1 +1,0 @@
-export const PLACE_HOLDER = '--';
