+++ frontend/src/components/SeedlotReviewSteps/Collection/utils.ts
@@ -1,5 +1,5 @@
+import { PLACE_HOLDER } from '../../../shared-constants/shared-constants';
 import { CodeDescResType } from '../../../types/CodeDescResType';
-import { PLACE_HOLDER } from './constants';
 
 /**
  * Format collection codes into human friendly words, e.g. [4, 5] -> "raking, picking".
