+++ frontend/src/shared-constants/shared-constants.ts
@@ -17,3 +17,5 @@ export const EmptyBooleanInputType: BooleanInputType = {
 export const SMALL_SCREEN_WIDTH = 320;
 export const MEDIUM_SCREEN_WIDTH = 672;
 export const LARGE_SCREEN_WIDTH = 1056;
+
+export const PLACE_HOLDER = '--';
