+++ oracle-api/src/main/java/ca/bc/gov/backendstartapi/endpoint/AreaOfUseEndpoint.java
@@ -1,6 +1,7 @@
 package ca.bc.gov.backendstartapi.endpoint;
 
 import ca.bc.gov.backendstartapi.config.SparLog;
+import ca.bc.gov.backendstartapi.dto.AreaOfUseDto;
 import ca.bc.gov.backendstartapi.dto.SpzDto;
 import ca.bc.gov.backendstartapi.entity.FundingSource;
 import ca.bc.gov.backendstartapi.security.RoleAccessConfig;
@@ -14,6 +15,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
+import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -63,5 +65,31 @@ public class AreaOfUseEndpoint {
     SparLog.info("Recevied request to fetch all SPZs for vegetation code: {}", vegCode);
 
     return areaOfUseService.getSpzByVegCode(vegCode.toUpperCase());
+  }
+
+  /**
+   * Get SPZ and SPU-geospatial information.
+   *
+   * @param spuId A seed plan unit id.
+   * @return A {@link SpzSpuGeoDto} containing the results.
+   */
+  @GetMapping(path = "/spu/{spuId}", produces = MediaType.APPLICATION_JSON_VALUE)
+  @Operation(
+      summary = "Get area of use information given an SPU id",
+      description = "Fetch geospatial data for the SPU along with a list of SPZ information.",
+      responses = {
+        @ApiResponse(
+            responseCode = "200",
+            description = "A data object containing information found."),
+        @ApiResponse(
+            responseCode = "401",
+            description = "Access token is missing or invalid",
+            content = @Content(schema = @Schema(implementation = Void.class)))
+      })
+  @RoleAccessConfig({"SPAR_TSC_ADMIN", "SPAR_MINISTRY_ORCHARD", "SPAR_NONMINISTRY_ORCHARD"})
+  public AreaOfUseDto getAreaOfUseData(
+      @Parameter(description = "The SPU (Seed Planning Unit) ID") @PathVariable("spuId")
+          Integer spuId) {
+    return areaOfUseService.calcAreaOfUseData(spuId);
   }
 }
